requirements.txt
openai-whisper
customtkinter
opencv-python
moviepy==1.0.3
